Revise code of listMgr() command PLAYLIST_ITEM_GET_POS.
Use uintptr_t type return variable.

Additionally, include stdint.h.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@36868 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ gui/util/list.c
@@ -21,6 +21,7 @@
  * @brief List management
  */
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -52,7 +53,7 @@ static urlItem *urlList;
  */
 void *listMgr(int cmd, void *data)
 {
-    unsigned int pos;
+    uintptr_t pos;
     plItem *pdat  = (plItem *)data;
     urlItem *udat = (urlItem *)data;
 
@@ -127,7 +128,7 @@ void *listMgr(int cmd, void *data)
         pos = 0;
 
         if (plList) {
-            unsigned int i = 0;
+            uintptr_t i  = 0;
             plItem *item   = plList;
 
             do {
